third_party/blink/web_tests/fast/forms/method-attribute-expected.txt

Test method and formmethod attributes

Missing value default:
PASS form1.method is "get"
PASS form1.getAttribute("method") is null


Invalid value default:
PASS form1.method = " post "; form1.method is "get"
PASS form1.getAttribute("method") is " post "
PASS form1.setAttribute("method", "put"); form1.method is "get"


Valid values:
PASS form1.method = "Post"; form1.method is "post"
PASS form1.getAttribute("method") is "Post"
PASS form1.method = "Get"; form1.method is "get"
PASS form1.getAttribute("method") is "Get"


PASS successfullyParsed is true

TEST COMPLETE
